Fix IO exceptions being swallowed
Error code 32 is actually an IO exception which happens when a file is
locked by another process. As the code is now, this exception is being
caught, its error code is then compared to some predermined results,
and as it's not equal to them it's being swallowed (no rethrows here).

This should actually be ==, not !=, see how @nblumhardt does it here:
https://github.com/serilog/serilog-dnx-prerelease/blob/master/src/
Serilog.Dnx.Prerelease/Sinks/RollingFile/RollingFileSink.cs#L150